requirements.txt
# LangChain LLM Chain Basics - Dependencies
# Install with: pip install -r requirements.txt

# LangChain framework
langchain>=0.1.0

# OpenAI integration for LangChain
langchain-openai>=0.0.5

# Environment variables management
python-dotenv>=1.0.0

# OpenAI Python client (dependency of langchain-openai)
openai>=1.0.0
